backend/requirements.txt
discord.py
slack-sansio[aiohttp]
PyYAML
aiohttp_cors
